+++ app/assets/javascripts/backend/head.js
@@ -16,7 +16,5 @@
 //= require bootstrap-switch
 
 // Ckeditor
-// require ckeditor/override
-//= require ckeditor/init
 //= require backend/mokio_dynamic_fields
 //= require backend/lodash/lodash